chore: Rename DDL to Uphoster
