+++ bot/modules/__init__.py
@@ -11,7 +11,7 @@ from .gd_search import gdrive_search, select_type
 from .help import arg_usage, bot_help
 from .mediainfo import mediainfo
 from .broadcast import broadcast
-from .ddl import ddl
+from .uphoster import uphoster
 from .mirror_leech import (
     mirror,
     leech,
@@ -63,7 +63,7 @@ __all__ = [
     "gdrive_search",
     "select_type",
     "arg_usage",
-    "ddl",
+    "uphoster",
     "mirror",
     "leech",
     "qb_leech",
